+++ ddbmock/routes/create_table.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from pyramid.view import view_config
 from ddbmock.database import DynamoDB
 from ddbmock.validators import dynamodb_api_validate
 from ddbmock.errors import wrap_exceptions
 
-# Real work
 @wrap_exceptions
 @dynamodb_api_validate
 def create_table(post):
@@ -17,8 +15,3 @@ def create_table(post):
     return {
         "TableDescription": desc,
     }
-
-# Pyramid route wrapper
-@view_config(route_name='create_table', renderer='json')
-def pyramid_create_table(request):
-    return create_table(request.json)