+++ ddbmock/routes/delete_item.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
 
-from pyramid.view import view_config
 from ddbmock.database import DynamoDB
 from ddbmock.validators import dynamodb_api_validate
 from ddbmock.errors import wrap_exceptions, ResourceNotFoundException
 
-# Real work
 @wrap_exceptions
 @dynamodb_api_validate
-def put_item(post):
+def delete_item(post):
     #FIXME: this line is a temp workaround
     if u'ReturnValues' not in post:
         post[u'ReturnValues'] = u"NONE"
@@ -17,19 +15,15 @@ def put_item(post):
 
     name = post[u'TableName']
     table = DynamoDB().get_table(name)
-    old, new = table.put(post[u'Item'], post[u'Expected'])
-    units = max(old.get_size().as_units(), new.get_size().as_units())
+    item = table.delete_item(post[u'Key'], post[u'Expected'])
 
     ret = {
-        "ConsumedCapacityUnits": units,
+        "ConsumedCapacityUnits": item.get_size().as_units(),
+        "Attributes": item,
     }
 
     if post[u'ReturnValues'] == "ALL_OLD":
-        ret["Attributes"] = old
-
-    return ret
-
-# Pyramid route wrapper
-@view_config(route_name='put_item', renderer='json')
-def pyramid_put_item(request):
-    return put_item(request.json)
+        return ret
+    else:
+        del ret["Attributes"]
+        return ret