+++ ddbmock/routes/delete_table.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from pyramid.view import view_config
 from ddbmock.database import DynamoDB
 from ddbmock.validators import dynamodb_api_validate
 from ddbmock.errors import wrap_exceptions
 
-# Real work
 @wrap_exceptions
 @dynamodb_api_validate
 def delete_table(post):
@@ -15,8 +13,3 @@ def delete_table(post):
     return {
         'TableDescription': ret,
     }
-
-# Pyramid route wrapper
-@view_config(route_name='delete_table', renderer='json')
-def pyramid_delete_table(request):
-    return delete_table(request.json)