+++ ddbmock/routes/get_item.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from pyramid.view import view_config
 from ddbmock.database import DynamoDB
 from ddbmock.validators import dynamodb_api_validate
 from ddbmock.errors import wrap_exceptions
 
-# Real work
 @wrap_exceptions
 @dynamodb_api_validate
 def get_item(post):
@@ -28,8 +26,3 @@ def get_item(post):
         return {
             "ConsumedCapacityUnits": base_capacity,
         }
-
-# Pyramid route wrapper
-@view_config(route_name='get_item', renderer='json')
-def pyramid_get_item(request):
-    return get_item(request.json)