+++ ddbmock/routes/scan.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from pyramid.view import view_config
 from ddbmock.database import DynamoDB
 from ddbmock.validators import dynamodb_api_validate
 from ddbmock.errors import wrap_exceptions, ResourceNotFoundException, ValidationException
 
-# Real work
 @wrap_exceptions
 @dynamodb_api_validate
 def scan(post):
@@ -22,7 +20,7 @@ def scan(post):
         post[u'ExclusiveStartKey'] = None
 
     if post[u'AttributesToGet'] and post[u'Count']:
-        raise ValidationException("Can filter fields when only count is requested")
+        raise ValidationException("Can not filter fields when only count is requested")
 
     name = post[u'TableName']
     table = DynamoDB().get_table(name)
@@ -45,8 +43,3 @@ def scan(post):
         ret[u'Items'] = results.items
 
     return ret
-
-# Pyramid route wrapper
-@view_config(route_name='scan', renderer='json')
-def pyramid_scan(request):
-    return scan(request.json)