+++ ddbmock/routes/update_table.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
-from pyramid.view import view_config
 from ddbmock.database import DynamoDB
 from ddbmock.validators import dynamodb_api_validate
 from ddbmock.errors import wrap_exceptions, ResourceNotFoundException
 
-# Real work
 @wrap_exceptions
 @dynamodb_api_validate
 def update_table(post):
@@ -22,8 +20,3 @@ def update_table(post):
     return {
         "TableDescription": desc,
     }
-
-# Pyramid route wrapper
-@view_config(route_name='update_table', renderer='json')
-def pyramid_update_table(request):
-    return update_table(request.json)